+++ FA2sp/Ext/CFinalSunDlg/Hooks.cpp
@@ -159,6 +159,7 @@ DEFINE_HOOK(43209D, CFinalSunDlg_Update_TranslateMenuItems, A)
     translateMenuItem(40092, "Menu.MapTools.PaintCliffFront");
     translateMenuItem(40093, "Menu.MapTools.PaintCliffBack");
     translateMenuItem(40133, "Menu.MapTools.SearchWaypoint");
+    translateMenuItem(40134, "Menu.MapTools.NavigateCoordinate");
     translateMenuItem(40135, "Menu.MapTools.ToolScripts");
 
     if (0)